Created Physics Asset for Darth Vader
I created a Physics Asset for the Darth Vader character and restricted the collisions to each other. But the scale of the bone is too big and I tried to edit it in Blender but it didn't work. Then I was able to successfully Retarget the UE5 model, but the root bone of the Darth Vader skeleton was in a very different place, even if the animations worked properly, the Root Motion was not working, so this goal was abandoned. When the time comes, I will change the bone myself if necessary and achieve this, but this will remain like this for now due to the waste of time.